# Improve TSemaphore